Commit All

+++ api/src/main/java/com/phone/api/web/rest/DistrictResource.java
@@ -1,7 +1,9 @@
 package com.phone.api.web.rest;
 
 import com.phone.api.domain.District;
+import com.phone.api.repository.CustomerRepository;
 import com.phone.api.repository.DistrictRepository;
+import com.phone.api.service.CustomerService;
 import com.phone.api.service.DistrictService;
 import com.phone.api.service.dto.DistrictDTO;
 import com.phone.api.utilty.HeaderUtil;
@@ -48,16 +50,20 @@ public class DistrictResource {
 
     private final Logger log = LoggerFactory.getLogger(DistrictResource.class);
     private static final String ENTITY_NAME = "DISTRICT";
+    private final CustomerRepository customerRepository;
 
     @Value("phoneListApp")
     private String applicationName;
 
     private final DistrictService districtService;
+    private final CustomerService customerService;
     private final DistrictRepository districtRepository;
 
-    public DistrictResource(DistrictService districtService, DistrictRepository districtRepository) {
+    public DistrictResource(DistrictService districtService, DistrictRepository districtRepository, CustomerService customerService, CustomerRepository customerRepository) {
         this.districtService = districtService;
         this.districtRepository = districtRepository;
+        this.customerService = customerService;
+        this.customerRepository = customerRepository;
     }
 
     /**
@@ -214,6 +220,12 @@ public class DistrictResource {
         log.debug("REST request to delete District : {}", id);
         if (!districtRepository.existsById(id)) {
             throw new ResourceNotFoundException("Not Found District with id = " + id);
+        }
+        if (customerRepository.findAllCustomersAmountThatUsingDistricts(id) > 0) {
+            throw new ResourceNotFoundException("Not Found Customers of District with id = " + id);
+            /**
+             * TODO : Burada kaldım. Count miktarının sıfırdan büyük olması sonucu bir hata fırlatması gerekir. Bundan dolayı contorller advice a yeni bir method eklememiz gerekmektedir.
+             */
         }
         districtService.delete(id);
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
